Add files via upload
Added the delete button for each item that is added to the shopping list as well as the option to select each item and unselect it by clicking on it. A function itemclicked was created and the line-through css decoration was used. The delete option is still not working properly as it is deleting items that are not supposed to be deleted. The adding to the list is also replacing any items already present.